r1a config upload
config is a RW OS that boots into xterm for testing. Uploading to capture the addition of the python-rpi-gpio library. I had a GPIO but not the right one for python, which is what was failing the script.

This config should be updated to NOT be RW and also boot into the script correct (via xinitrc file).